Add i18n module
This will hold translation strings. It is intended to use different files for different aspects of the application. If we can match the module boundaries is questionable. I assume that might be too fine grained.
That module has no dependency to UI so can be used also for error logging in headless context, etc.