Remove reference to common-platform
There are no lombok dependency version contraints defined in common-platform,
so lombok-dependencies.gradle does not need to consume that platform.

In the normal use case, lombok dependency version contraints would be defined
in common-platform, and common-platform would be consumed by
lombok-dependencies.gradle.  However, this does not work for the gradle
dependency configuration types `annotationProcessor` and `testAnnotationProcessor`
(could be a Gradle bug).  In this corner case, we do have to specify the version
as @chimp1984 has done in lombok-dependencies.gradle.