Use NonCachingController for wallet tab controllers which fill the lists at onViewAttached.
The tabs are cached by default so the onViewAttached is only called the first time
and again when navigating from another view to the parent tabview.
The NonCachingController avoids the caching.

But we should probably use a different strategy as we do not want to reload each
time the user navigates between tabs.
An option would be to keep a ObserveableSet in the walletService and trigger a
rpc call for reloading the data once the wallet state has changed.
To get pushed relevant wallet state changes would avoid requirement for
pulling. I guess with ZeroMQ that should work. Otherwise if there is an equivalent to
blocknotify for wallet changes we could use that.