Rename AuthenticatedDataImpl to DefaultAuthenticatedData
Remove data storage tests as most are not working anymore with the protobuf dependency and outdated anyway.
Use cases are tested with the app features and seems ok so far.
Still would be good to get test added at some point, but we need to find a solution how to deal with mocks
for protobuf definitions.