375-I2P-Improvements
Initial commit for changes related to improvements on I2P layer
Improvements to the I2P Layer + Implementation of log mechanism for I2P local router (enabled by config variable)
Extracting I2pRouter to it's own class and making a singleton out of it.
From now on, whenever using the embedded router, only one router will be started.
Making the client always connect to the configured port for the external router
Adding code to check if port is in use.

Dev: @c4talys7
Signed-off-by: c4talys7 <c4talys7@protonmail.com>