Add ability to check whether a given onion service is online
When another peer connects to us over Tor the connecting peer can lie
about its onion address. We can challenge the peer to sign a payload to
proof that the connecting peer has the private key of the provided onion
address. This change adds the ability to check whether the onion service
is known to the Tor directory servers.