Introduce getter for selectedUserIdentityObservable
As discussed in https://github.com/bisq-network/bisq2/discussions/835
this is an adaption to the suggested scheme.

In this case it's expected that the Observed value is always not null,
which can be seen in how different parts of the code handle the null
case. Having @Nullable annotation on the getter will help going forward.