Basic testing of MaterialPasswordField
This commit introduces the TestFX framework which is the standard for
JavaFX testing. Added a simple test that pulls up a
MaterialPasswordField and verifies that contents are properly passed to
the password field and that the text is masked when the eye icon is
clicked. This should serve as a baseline for #895.