Use a random default keyId
The default networkId is using the default key ID. It usually is not used for sending private mailbox messages, but in case it would we would have used the "default" value for it and all nodes would try to decrypt the encrypted message (by looking up the key with that id).
To avoid that, we use now a random keyId prefixed with "default_" created at first start and persisted.