Adjusting tab bug when animations are off
There were two issues affecting the rendering of the green tab markers lines. One of them was related to the maybeAnimateMark function only being called when transitionStarted was true. After removing the condition, rendering was breaking when the user view was the first to be opened, and that was adjusted increasing the timing of the transition painting method by 150ms

bisq-network/bisq2#1210