Change TabView menu to an always visible position
To achieve a better user experience, the menu in tab view
is modified to an always visible position. Now, only content is
scrollable. In this way, it is possible to jump to other tabs
as well as know where you are at any position.

This change is also necessary to implement the new design
for the notification panel.

Towards #1211