Use the randomly created key ID for the default Key ID instead of "default".
For mailbox messages we use the key ID to find our key pair. Using default would trigger a match for all users but decryption would fail.
The KeyPairStore creates a random default key ID at first usage and persists that for later access, so the key ID stays the same as long the KeyPairStore is not deleted.