Add config for passing the tor private key for the default node.
We do ot read the private key anymore from the tor directory but only write it as backup. We also do not write anymore the open SSH format as that is not used and it should avoid giving the impression that key is generated by tor. We also reduce the support to the default node as that is the only use case for now.