Replace contract for offer when creating a trade
Right after a peer takes an offer, now it is possible to send account
details directly without having to get the response from the maker.
Also, since we don't know the order in which the messages will be processed
at the maker side, we allow to create a trade without the contract (but
offer instead since that is always guaranteed to be there).
Contract is then set when BisqEasyTakeOfferEvent is handled.