Add ExcludeForHash annotation Add getBuilder method to Proto Add serializeForHash method
When serializeForHash is used we clear all fields annotated with `ExcludeForHash`.
This allows more flexibility for not identity-defining fields in use cases where we use the hash like in: data storage, pow, signatures
The getBuilder default method is just temporary to avoid the need to update all implementations of Proto.
Follow up commits will apply the new interface to the code base