Do not support nullable selectedUserIdentity
The only valid case where we could set it to null was at delete, but there we set then the next item from the userIdentities and we do not allow to delete the last item in userIdentities. Thus, we always have a selectedUserIdentity. Only at first startup before we create our user identity it would be not set, but the domain is not used before that moment.