build-logic: Support m out of n signature verification
In Electrum all developers sign the binaries, but in Bitcoin Core a
subset of its developers sign the binaries. Our existing signature
verification code fails if not all developers have signed the
binaries.

This change enforces that all provided signatures are signed by its
developers, i.e., if we have 20 hard-coded public keys and a binary
has 10 signatures we ensure that the public keys of those 10 signatures
belong to the hard-coded developers and all signatures are valid.