Extract the offer list to its own MVC component.
The BisqEasyOfferbookView is very overloaded and hard to manage.
We should also extract the left markets panel in a similar style.

As some data is required by the main container we let the main model hold the data and pass them to the OfferList.
It also moved the cell factories from the util class to the view, so that it is all in one place.