Add HashCashV2TokenService
We used the serialized message for the payload in the ProofOfWork data, which can be quite large in case of inventory requests or other large messages.

To avoid that we introduce the new HashCashV2TokenService which use the hash of the serialized message instead, thus reducing the size of the payload field to 20 bytes.

To support backward compatability we add a new Feature for that and use the new version only if the peer support that feature.