Add messages which cannot be used to create a new channel to a pendingMessages set and reprocess it at next message.
This covers the case when the maker is offline while the taker/buyer takes the offer and sends the btc address. The maker once going online receives both as mailbox message without guaranteed order, and if the btc address message was processed first, it is dropped as it cannot be used to create a new channel.