If currency cannot be derived from locale we fall back to USD
This can happen for instance for `es_419` (Generic Latin America) or others.

Use builder for locale as prev. constructor is deprecated since Java 19.
Add log

This PR is based on https://github.com/bisq-network/bisq2/pull/3411 which could not get merged due github not recognizing the commit signature as valid (it shows a green mark though)