Use MaterialTextArea which has everything we want here.
Use our validator framework.

Remove char counter as it adds the requirement for the listener which seems not justified for that use case.
The validator is enough here to give feedback if the text is too long.

For builder style interface don't use `set` prefix.
Use value as param name to reduce redundancy and inconsistencies in naming.

Remove result handler and return TextInputDialog instead. A client can set their own result handler if they wish. Alternatively we could pass a consumer, but that seems not needed.

Remove show method as it has different signature as the base classes show method and is not used anyway.