Fix missing persistence for error messages
Persist was only called for network messages and user events but not for an FsmErrorEvent as that was triggered by the Fsm directly.
To ensure persist calls cannot be forgotten we add a persist() method to the handler base classes to ensure that persist is called after each commit method. There might be some un-needed persist calls in case there was no data committed.

For Bisq Easy this is not yet enforced, but will be handled in upcoming commits.