Remove unit tests for account and account payload.
The unit test are only testing protobuf but no logic which would benefit from tests (as there is none). With the ongoing refactorings it becomes very burdensome to maintain and adjust the tests without any visible benefit of the tests. Further there are many more accounts to be added and adding tests for all those would increase quite a lot dev effort, again with not much visible benefit.