Accommodate RPC API changes part 1: server-side buyer secret redaction
Make the buyer call 'Musig.GetPartialSignatures' a second time, when he
is ready to reveal his 'swapTxInputPartialSignature' upon payment start,
so that it no longer needs to be held on the client side, and is only
revealed by the server when the 'buyerReadyToRelease' flag is enabled.
(The second call is necessary anyway because the sighash-related changes
mean that the buyer's Rust server learns the partial signature a round
too late to include in the first response.)

Also enable 'sellerReadyToRelease' in the seller's 'Musig.SignSwapTx'
call upon receipt of the buyer's 'PaymentInitiatedMessage_E', as a quick
fix to prevent the Rust server from redacting the seller secrets (namely
the signed swap tx & the buyer-payout private key share) from the
response. These are no longer expected to be held on the Java side.