Use a thread per send.
Before we had a thread for the while broadcast, thus if one sending would have been slow it could have triggered a timeout and it was executed in sequence. The send call is blocking and we have throttling in the connection layer, which can pause that thread for about 1 seconds.

The broadcast and reBroadcast methods are called from dispatcher thread (single thread), thus we need to ensure we do not have any IO/blocking code in the execution path.