Add throttle to avoid network stress at republishing proofOfBurn and bondedReputation data.
Currently, as July 2025 we have about 150 messages we republish at startup.
With 4 seconds pause we get about 10 minutes until all data is published.

Alternatively we could check for data in the network DB and only republish those which are not present or close to expire date. But that would require also some extra republish handling as we cannot be sure that our own data in the network DB are representative for other nodes. Therefor, that change appears more simple and as restart of oracle nodes will become less frequent once the threading issues is fixed (should be fixed by now), the case where that is relevant will more rare.

Furthermore there are planned changes for the bridge API to change to grpc with data push, which will cause an overall rewrite of that service.