Change DISPATCHER executor to a pool with core pool size of 1 and max size of 20. As it uses a SynchronousQueue there is no queuing used.
We use a thread for each listener call in the network layer.
As it is expected that handlers do not do heavy or blocking work those tasks should complete very fast, making the thread idle again for re-use.
We want to avoid that any listener which runs into a waiting or blocking state, can stall other listeners or the system. Only if we would grow to 20 blocking threads the pool would reject any new task.

With setting the core pool size very low, we add some delay for creating new threads. If we set it higher the pool would create more core threads and consume more resources. This difference in behaviour have been observed when trying out different pool settings.