Add new wallet module
We will remove the wallets module in a later commit.
We copied over the most relevant classes, but they might change to the actual needs for the BDK wallet.

The original wallets module was designed with support for multiple wallets in mind. But as we use now BDK which has support for different blockchain backends, we do not need that flexibility anymore. The dependencies to the bitcoind project made a refactoring hard, thus it appeared easier to create a new simple module and remove the old one.