Ensure that the removal from the map happens before the requestFuture completes.
It might be that client handlers look up that map and the various whenComplete handlers have no defined order of execution. This can lead to the case that the client whenComplete is called before our class internal whenComplete is called where we remove the entry from the map.

This case happened in the InventoryService.

By adding the internal priorityFuture, we ensure that the removal from the map happens always before the requestFuture completes.

We also moved the timout handling into the RequestFuture to control the order of the priorityFuture completion.