Improve inventory request handling and policy.
When testing with I2P with some nodes which do not deliver the real network data (only a small subset, as those nodes are not up to date), the initial inventory request terminated incorrectly when it received the first response with maxSizeReached=false as well if entries have been empty (in which case we also have maxSizeReached=false).
We had incorrectly interpreted that as all data delivered and stopped requests. So clients did not receive the network data which requires multiple requests with the 2 MB cap.

Furthermore, we considered it complete when one peer response with all data delivered. This might be a valid case but if that node has a subset of the network data we would be also restricted to that subset. To increase chances that we get the full network view, we require now at least 3 (defined by minCompletedRequests) peers who have reported maxSizeReached=false and have delivered a non-empty inventory list.

Beside those fixed there have been several other improvements which should make the data delivery even in very imperfect environments more reliable. As said the I2P layer has atm several "broken" nodes which helped to surface those issues. With a larger and more robust node network those issues would not have become visible (as with Tor).