Exchange additional Claim Tx fields with peer
Add Claim Tx-related fields to 'PartialSignatures' and 'NonceShares' in
'trade.proto', so that they can be exchanged with the peer in addition
to the fields for the existing prepared (Swap, Warning & Redirect) txs.
Also make sure to include them in the 'PartialSignaturesMessage' and
'NonceSharesMessage' proto DTOs, for talking with the gRPC server.

This prevents warning logs on the Rust side that the Claim Txs cannot be
built and/or signed. (These will eventually be made hard errors.)