Pick up redirectionAmountMsat field from the MuSig server
Make sure to include the 'NonceSharesMessage.redirectionAmountMsat'
field, that was added to 'rpc.proto', in the corresponding Java DTO.
This is picked up from the Rust server in its 'GetNonceShares' response,
but not passed to the peer (so we don't add it to the corresponding
'NonceShares' message in 'trade.proto'). Instead, it is used to compute
(on the Java side) the absolute receiver amounts from the BM share list
picked up from the oracle. So pass the quantity as an argument to
'PartialSignaturesRequestUtil.getBurningManDPTReceivers'. (For now,
continue to return a static, dummy list of receivers and just log a
warning that the 'redirectionAmountMsat' argument was ignored.)

Also change the hardcoded Signet BM dummy addresses to Regtest ones, to
prevent wrong-network warnings on the Rust side (that will also be made
into hard errors eventually).