Specify a trade fee recipient in NonceSharesRequest
Populate the new 'NonceSharesRequest.tradeFeeReceiver' proto field with
a dummy recipient (BM address + amount) when calling 'GetNonceShares',
making the Rust server add an extra trade fee output to the Deposit Tx.
As commented in 'NonceSharesRequestUtil', this is always paid by seller,
so adjustments would need to be made to the requested trade amount and
buyer/seller security deposits to share the fees differently.