Add lookup map for maxMapSize by class name.
Improve prune method.

We want to avoid to take the maxMapSize from data we received from the network, as that data can have an older version of maxMapSize and is basically untrusted data.

If we create the MetaData by constructor we have the metadata set from our code, thus we do not need a lookup but put the value into the lookup map.