Add getMutableClone and use that in StorageService
This is only temporary to not risk that we get an exception if a client mutates the map which is a not valid case, but we the mobile release we prefer to avoid the risk to run into an exception.
For main and for a major mobile update we should remove that.