Fix minor flaws in receiver amounts calculation & add tests
Change the receiver amounts calculation slightly, to fix a couple of
minor flaws discovered while porting the algorithm to Rust (in order to
calculate precise fees for the Redirect Tx in the original end-to-end
protocol implementation, as part of PR 94 of 'bisq-musig'):

1) The intermediate floating-point receiver amounts should be rounded to
   nearest by 'Math::round', instead of down by casting to long, to
   eliminate bias, so that cumulative rounding errors grow at the square
   root of the output number, not linearly;

2) The fee-dependent 'minOutputAmount' should be rounded _up_ to the
   next whole satoshi value, not down though integer division by 1000.

Applying (1) potentially introduces a new problem, that in some edge
cases too many outputs get included to pay each receiver the minimum
allowed amount (say, if all the BM have equal shares paying just below
the minimum, which then get intermediately rounded up to the minimum,
causing all to get included and leading to total overpayment). So add
some logic to prevent that.

Also, backport unit tests written for the Rust version which exercise
those three issues in particular, and include a test for the associated
helper method, 'DelayedPayoutTxReceiverService::scriptPubKeyLength'.