Add I2P_ACTIVATIONE_DATE and consider AuthorizedOracleNode, AuthorizedBondedRole or AuthorizedAlertData invalid if they contain an I2P address.
We had oracle and seed nodes registered which had already I2P addresses defined, but later unregistered/banned those nodes.
As the 2.1.7 versions have an incorrect address validation for I2P the inventory response fail if it contains such data.
The removal of the nodes did not reach the whole network due the same issue with 2.1.7, thus we have such dangling data in the network.

Other I2P data is not expected in the network (e.g. user profiles, offers).